Compiled 3 files into 1 to make HL7Parser ultra-portable.
