--- HL7Segment.py
@@ -1,6 +1,0 @@
-class HL7Segment:
-    """HL7 Segment object."""
-
-    def __init__(self):
-        self.values = []
-        self.name = ''
